Removing unit test file for this example
JUnit tests are not part of the assignment description and are not
an especially useful tool for students who would be working on an
assignment like this. Useful tools might instead include standard
input values for the scanner to process and verify the code is
providing correct output, or at the very least, is not raising
uncaught exceptions.